requirement.txt
pandas 
numpy
seaborn
matplotlib
scikit-learn
xgboost
ipykernel
sqlalchemy
mysql-connector-python
ipython-sql
pyyaml
imblearn
joblib
